remove test_levo_integration.py
